Add files via upload
#include<iostream>:-This line includes the standard input/output library in the program. It allows you to use input and output functions like cin and cout.
using namespace std means that we can use names for objects and variables from the standard library.
int main():-This is the main function where the program execution begins.
This program is designed to take an integer input from the user and display it. 
An integer variable number is declared.
This program is used to demonstrate how to take an integer input from the user and display it.