Add files via upload
#include<iostream>:-This line includes the standard input/output library in the program. It allows you to use input and output functions like cin and cout.
using namespace std means that we can use names for objects and variables from the standard library.
int main():-This is the main function where the program execution begins.
This program is used to add two numbers entered by the user and display the result. 
Three integer variables are declared: first_number, second_number, and sum.
The program asks the user to enter two numbers.
The program adds the two numbers and stores the result in the sum variable.
The program outputs the sum of the two numbers.