Add files via upload
#include<iostream>:-This line includes the standard input/output library in the program. It allows you to use input and output functions like cin and cout.
using namespace std means that we can use names for objects and variables from the standard library.
int main():-This is the main function where the program execution begins.
This program demonstrates the use of relational operators to compare two integers.
Two integer variables a and b are declared and initialized with values 6 and 4, respectively.
Equality operator (==): Checks if a is equal to b.
Greater than operator (>): Checks if a is greater than b.
Greater than or equal to operator (>=): Checks if a is greater than or equal to b.
Less than operator (<): Checks if a is less than b.
Less than or equal to operator (<=): Checks if a is less than or equal to b.
Not equal to operator (!=): Checks if a is not equal to b.
This program is used to demonstrate how relational operators work. 
Relational operators are used to compare values and used in decision-making.