Add files via upload
#include<iostream>:-This line includes the standard input/output library in the program. It allows you to use input and output functions like cin and cout.
using namespace std means that we can use names for objects and variables from the standard library.
int main():-This is the main function where the program execution begins.
This program is designed to find the greatest of two numbers using the ternary (conditional) operator.
Two integer variables a and b are declared and initialized with values 3 and 4, respectively.
The program uses the ternary operator to compare a and b. 
If a is less than b, result is assigned the value of b; otherwise, it is assigned the value of a.
This program is used to determine and display the greater of two integers. 
It demonstrates the use of the ternary operator for conditional assignments.